requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - SQLite (built into Python)
sqlalchemy==2.0.23

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI/ML - Multiple Providers
openai==1.3.8
anthropic==0.25.1
google-generativeai==0.5.4
tiktoken==0.5.2

# HTTP Client for AI APIs
httpx==0.25.2
aiohttp==3.9.3
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy>=1.24.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.6.0

# Logging
structlog==23.2.0

# CORS
fastapi-cors==0.0.6

# Email (Optional)
fastapi-mail==1.4.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Validation
email-validator==2.1.0

# Date/Time
python-dateutil==2.8.2

# Utils
click==8.1.7
rich==13.7.0

# Additional dependencies for AI Poker Coach
aiofiles==23.2.1
Pillow==10.1.0
pydantic-extra-types==2.2.0
tenacity==8.2.3